Refactor item tests
